added backend register user

+++ emswebapp/src/main/java/com/cosmostaker/emswebapp/controller/EmployeeController.java
@@ -61,7 +61,7 @@ public class EmployeeController {
 
     // Mapping for update employee form
     @GetMapping("/showUpdateEmployeeForm{id}")
-    public String showUpdateEmployeeForm(@PathVariable(value = "id") int id, Model model) {
+    public String showUpdateEmployeeForm(@PathVariable(value = "id") Long id, Model model) {
         // get employee from the service
         Employee employee = employeeService.getEmployeeById(id);
 
@@ -76,7 +76,7 @@ public class EmployeeController {
 
     // Delete employee
     @GetMapping("/deleteEmployee{id}")
-    public String deleteEmployee(@PathVariable(value = "id") int id) {
+    public String deleteEmployee(@PathVariable(value = "id") Long id) {
         // call delete employee method
         this.employeeService.deleteEmployeeById(id);
